Regim/requirements.txt
SimpleITK==1.1.0
pydicom==1.1.0
pypng==0.0.18
numpy==1.15.0
opencv-contrib-python==3.4.2.17
Pillow==5.2.0